backend/pyproject.toml
[project]
name = "coach-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.5.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]
